addressing feedback - wip
