+++ src/NuGet.Services.AzureSearch/Models/SearchDocument.cs
@@ -45,6 +45,22 @@ namespace NuGet.Services.AzureSearch
             public string FullVersion { get; set; }
             public string[] Versions { get; set; }
             public string[] PackageTypes { get; set; }
+
+            /// <summary>
+            /// The list of a package's supported target framework generations. These framework generations will be standardized
+            /// strings of the form "net", "netframework", "netcore", or "netstandard", for the four generations supported
+            /// by the NuGet.org filtering experience.
+            /// </summary>
+            [SimpleField(IsFilterable = true)]
+            public string[] Frameworks { get; set; }
+
+            /// <summary>
+            /// The list of a package's supported target framework monikers, stored as their normalized TFM strings 
+            /// (same as the 'short folder name'). eg. net472, netcoreapp3.1
+            /// </summary>
+            [SimpleField(IsFilterable = true)]
+            public string[] Tfms { get; set; }
+
             public bool? IsLatestStable { get; set; }
             public bool? IsLatest { get; set; }
         }
