+++ tests/NuGet.Services.AzureSearch.Tests/SearchDocumentBuilderFacts.cs
@@ -392,6 +392,8 @@ namespace NuGet.Services.AzureSearch
       ""packageTypes"": [
         ""Dependency""
       ],
+      ""frameworks"": [],
+      ""tfms"": [],
       ""isLatestStable"": false,
       ""isLatest"": true,
       ""semVerLevel"": 2,
@@ -401,7 +403,6 @@ namespace NuGet.Services.AzureSearch
       ""description"": ""Description."",
       ""fileSize"": 3039254,
       ""flattenedDependencies"": ""Microsoft.Data.OData:5.6.4:net40-client|Newtonsoft.Json:6.0.8:net40-client"",
-      ""frameworks"": null,
       ""hash"": ""oMs9XKzRTsbnIpITcqZ5XAv1h2z6oyJ33\u002BZ/PJx36iVikge/8wm5AORqAv7soKND3v5/0QWW9PQ0ktQuQu9aQQ=="",
       ""hashAlgorithm"": ""SHA512"",
       ""iconUrl"": ""http://go.microsoft.com/fwlink/?LinkID=288890"",
@@ -430,7 +431,6 @@ namespace NuGet.Services.AzureSearch
         ""Scalable"",
         ""windowsazureofficial""
       ],
-      ""tfms"": null,
       ""title"": ""Windows Azure Storage"",
       ""tokenizedPackageId"": ""WindowsAzure.Storage"",
       ""lastCommitTimestamp"": ""2018-12-13T12:30:00\u002B00:00"",
@@ -721,6 +721,12 @@ namespace NuGet.Services.AzureSearch
       ""packageTypes"": [
         ""Dependency""
       ],
+      ""frameworks"": [
+        ""netframework""
+      ],
+      ""tfms"": [
+        ""net40-client""
+      ],
       ""isLatestStable"": false,
       ""isLatest"": true,
       ""semVerLevel"": 2,
@@ -730,9 +736,6 @@ namespace NuGet.Services.AzureSearch
       ""description"": ""Description."",
       ""fileSize"": 3039254,
       ""flattenedDependencies"": ""Microsoft.Data.OData:5.6.4:net40-client|Newtonsoft.Json:6.0.8:net40-client"",
-      ""frameworks"": [
-        ""netframework""
-      ],
       ""hash"": ""oMs9XKzRTsbnIpITcqZ5XAv1h2z6oyJ33\u002BZ/PJx36iVikge/8wm5AORqAv7soKND3v5/0QWW9PQ0ktQuQu9aQQ=="",
       ""hashAlgorithm"": ""SHA512"",
       ""iconUrl"": ""http://go.microsoft.com/fwlink/?LinkID=288890"",
@@ -760,9 +763,6 @@ namespace NuGet.Services.AzureSearch
         ""Queue"",
         ""Scalable"",
         ""windowsazureofficial""
-      ],
-      ""tfms"": [
-        ""net40""
       ],
       ""title"": ""Windows Azure Storage"",
       ""tokenizedPackageId"": ""WindowsAzure.Storage"",
@@ -866,6 +866,94 @@ namespace NuGet.Services.AzureSearch
 
                 Assert.Equal(Data.GalleryLicenseUrl, document.LicenseUrl);
             }
+
+            [Theory]
+            [MemberData(nameof(TargetFrameworkCases))]
+            public void AddsFrameworksAndTfmsFromPackage(List<string> supportedFrameworks, List<string> expectedTfms, List<string> expectedFrameworks)
+            {
+                // arrange
+                var package = new Package
+                {
+                    PackageRegistration = new PackageRegistration
+                    {
+                        Id = "TestPackage",
+                    },
+                    Id = "TestPackage",
+                    NormalizedVersion = Data.NormalizedVersion,
+                    LicenseExpression = "Unlicense",
+                    HasEmbeddedIcon = true,
+                    SupportedFrameworks = supportedFrameworks
+                                                .Select(f => new PackageFramework() { TargetFramework = f })
+                                                .ToArray(),
+                };
+
+                // act
+                var document = _target.FullFromDb(
+                    Data.PackageId,
+                    Data.SearchFilters,
+                    Data.Versions,
+                    isLatestStable: false,
+                    isLatest: true,
+                    fullVersion: Data.FullVersion,
+                    package: package,
+                    owners: Data.Owners,
+                    totalDownloadCount: Data.TotalDownloadCount,
+                    isExcludedByDefault: false);
+
+                // assert
+                Assert.True(document.Tfms.Length == expectedTfms.Count);
+                foreach (var item in expectedTfms)
+                {
+                    Assert.Contains(item, document.Tfms);
+                }
+
+                Assert.True(document.Frameworks.Length == expectedFrameworks.Count);
+                foreach (var item in expectedFrameworks)
+                {
+                    Assert.Contains(item, document.Frameworks);
+                }
+            }
+
+            public static IEnumerable<object[]> TargetFrameworkCases =>
+                new List<object[]>
+                {
+                    new object[] {new List<string> {}, new List<string>(), new List<string> {}},
+                    new object[] {new List<string> {"any"}, new List<string> {"any"}, new List<string> {}},
+                    new object[] {new List<string> {"net"}, new List<string> {"net"}, new List<string> {"netframework"}},
+                    new object[] {new List<string> {"win"}, new List<string> {"win"}, new List<string> {}},
+                    new object[] {new List<string> {"foo"}, new List<string> {}, new List<string> {}}, // unsupported tfm is not included
+                    new object[] {new List<string> {"dotnet"}, new List<string> {"dotnet"}, new List<string> {}},
+                    new object[] {new List<string> {"net472"}, new List<string> {"net472"}, new List<string> {"netframework"}},
+                    new object[] {new List<string> {"net40-client"}, new List<string> {"net40-client"}, new List<string> {"netframework"}},
+                    new object[] {new List<string> {"net5.0"}, new List<string> {"net5.0"}, new List<string> {"net"}},
+                    new object[] {new List<string> {"netcoreapp3.0"}, new List<string> { "netcoreapp3.0" }, new List<string> { "netcore" } },
+                    new object[] {new List<string> {"netstandard2.0"}, new List<string> { "netstandard2.0" }, new List<string> { "netstandard" } },
+                    new object[] {new List<string> {"net40", "net45"}, new List<string> {"net40", "net45"}, new List<string> {"netframework"}},
+                    new object[] {new List<string> {"net5.0-tvos", "net5.0-ios"}, new List<string> {"net5.0-ios", "net5.0-tvos"}, new List<string> {"net"}},
+                    new object[] {new List<string> {"net5.0-tvos", "net5.0-ios13.0"}, new List<string> {"net5.0-ios13.0", "net5.0-tvos"}, new List<string> {"net"}},
+                    new object[] {new List<string> {"net5.1-tvos", "net5.1", "net5.0-tvos"}, new List<string> {"net5.0-tvos", "net5.1", "net5.1-tvos"}, new List<string> {"net"}},
+                    new object[] {new List<string> {"net5.0", "netcoreapp3.1", "native"}, new List<string> {"native", "net5.0", "netcoreapp3.1"}, new List<string> {"net", "netcore"}},
+
+                    new object[] {new List<string> {"netcoreapp3.1", "netstandard2.0"}, new List<string> {"netcoreapp3.1", "netstandard2.0"},
+                                    new List<string> {"netcore", "netstandard"}},
+
+                    new object[] {new List<string> {"netstandard2.1", "net45", "net472", "tizen40"}, new List<string> {"netstandard2.1", "net45", "net472", "tizen40"},
+                                    new List<string> {"netframework", "netstandard"}},
+
+                    new object[] {new List<string>{"net40", "net471", "net5.0-watchos", "netstandard2.0", "netstandard2.1"},
+                                    new List<string>{"net40", "net471", "net5.0-watchos", "netstandard2.0", "netstandard2.1"}, new List<string> {"netframework", "net", "netstandard"}},
+
+                    new object[] {new List<string>{"net45", "netstandard2.1", "xamarinios"}, new List<string>{"net45", "netstandard2.1", "xamarinios"},
+                                    new List<string> {"netframework", "netstandard"}},
+
+                    new object[] {new List<string> {"portable-net45+sl4+win+wp71", "portable-net45+sl5+win+wp71+wp8"},
+                                    new List<string> {"portable-net45+sl4+win+wp71", "portable-net45+sl5+win+wp71+wp8"}, new List<string> {}},
+
+                    new object[] {new List<string> {"net20", "net35", "net40", "net45", "netstandard1.0", "netstandard1.3", "netstandard2.0",
+                            "portable-net40+sl5+win8+wp8+wpa81", "portable-net45+win8+wp8+wpa81"},
+                                    new List<string> {"net20", "net35", "net40", "net45", "netstandard1.0", "netstandard1.3", "netstandard2.0",
+                            "portable-net40+sl5+win8+wp8+wpa81", "portable-net45+win8+wp8+wpa81"}, new List<string> {"netframework", "netstandard"}}
+                };
         }
 
         public abstract class BaseFacts
